requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
absl-py==0.12.0
aiohttp==3.7.4.post0
antlr4-python3-runtime==4.8
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
black==20.8b1
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
configparser==5.0.2
cycler==0.10.0
decorator==4.4.2
dill==0.3.3
docker-pycreds==0.4.0
e2cnn==0.1.7
flake8==3.9.0
fsspec==0.8.7
future==0.18.2
gitdb==4.0.7
GitPython==3.1.14
google-auth==1.28.0
google-auth-oauthlib==0.4.3
grpcio==1.36.1
h5py==3.2.1
hydra-core==1.1.0.dev3
hydra-submitit-launcher==1.1.0
idna==2.10
iniconfig==1.1.1
ipykernel==5.5.0
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.0
jupyter-client==6.1.11
jupyter-core==4.7.1
kiwisolver==1.3.1
lie-conv==0.1
lightning-bolts==0.3.2
Markdown==3.3.4
matplotlib==3.3.4
mccabe==0.6.1
more-itertools==8.7.0
multidict==5.1.0
mypy-extensions==0.4.3
natsort==7.1.1
numpy==1.20.1
oauthlib==3.1.0
olive-oil-ml==0.1
omegaconf==2.1.0.dev24
packaging==20.9
pandas==1.2.2
parso==0.8.1
pathspec==0.8.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pip==20.3.3
pluggy==0.13.1
promise==2.3
prompt-toolkit==3.0.16
protobuf==3.15.6
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.8.0
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
pytorch-lightning==1.2.7
pytz==2021.1
PyYAML==5.3.1
pyzmq==22.0.3
regex==2021.3.17
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.0
sentry-sdk==1.0.0
setuptools==52.0.0.post20210125
shortuuid==1.0.1
six==1.15.0
sklearn==0.0
smmap==4.0.0
submitit==1.3.0
subprocess32==3.5.4
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.1.0
toml==0.10.2
torch==1.8.1+cu111
torchcontrib==0.0.2
torchdiffeq==0.2.1
torchmetrics==0.2.0
torchvision==0.9.1+cu111
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
wandb==0.10.25
wcwidth==0.2.5
Werkzeug==1.0.1
wheel==0.36.2
yarl==1.6.3
